refactor: profile 하위 폴더 파일명 표준화
