+++ app/profile/data_deletion.tsx
@@ -1,4 +1,4 @@
-import DataDeletionScreen from '@/src/screens/profile/data_deletion/data_deletion_screen';
+import DataDeletionScreen from '@/src/screens/profile/data_deletion';
 
 export default function DataDeletionRoute() {
   return <DataDeletionScreen />;
