+++ app/profile/edit.tsx
@@ -1,4 +1,4 @@
-import EditProfileScreen from '@/src/screens/profile/edit/edit_profile_screen';
+import EditProfileScreen from '@/src/screens/profile/edit';
 
 export default function EditProfileRoute() {
   return <EditProfileScreen />;
